backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic-settings==2.10.1
skyfield==1.53
tzdata==2025.2
timezonefinder==8.2.1
